Use release name + activemq-artemis for the secret env template
This clears up the issue of what scope the template is being evaluated
in when it is used from both the activemq-artemis chart and the
artemiq-application (parent) chart.

Before this fix, that template would render different secret names
depending on the chart scope.

Hardcoding the chart name into the template seems like the easiest
solution and isn't all that bad seeing as how the template itself
is also named the same way